norme step 4
