+++ includes/ft_ls_m.h
@@ -6,7 +6,7 @@
 /*   By: wzei <wzei@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/22 19:44:32 by wzei              #+#    #+#             */
-/*   Updated: 2019/10/23 13:55:40 by wzei             ###   ########.fr       */
+/*   Updated: 2019/10/23 15:02:06 by wzei             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -58,6 +58,12 @@ void					print_list(t_mlist *node, t_max_out *max,
 										char *dir_name);
 void					print_long_entry(t_fileinfo *file,
 										t_max_out *max, char *path);
+void					print_reg_dir(t_long_out *l_o,
+										t_fileinfo *file, t_max_out *max);
+void					print_link(t_long_out *l_o, t_fileinfo *file,
+										char *path, t_max_out *max);
+void					print_device(t_long_out *l_o, t_fileinfo *file,
+										t_max_out *max);
 
 /*
 ** Error handling
