+++ srcs/traverce_cur_dir.c
@@ -6,7 +6,7 @@
 /*   By: wzei <wzei@student.42.fr>                  +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2019/10/20 14:22:18 by wzei              #+#    #+#             */
-/*   Updated: 2019/10/23 13:38:05 by wzei             ###   ########.fr       */
+/*   Updated: 2019/10/23 14:12:25 by wzei             ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,16 +14,6 @@
 
 int			g_r;
 const char	*g_program_name = "ft_ls";
-
-/*
-**static void	init_lst(t_mlist *lst)
-**{
-**	lst->next = NULL;
-**	lst->prev = NULL;
-**	lst->content = NULL;
-**	lst->tag = 0;
-**}
-*/
 
 static void	print_head(const char *filename, t_max_out *max)
 {
@@ -41,31 +31,6 @@ static void	print_head(const char *filename, t_max_out *max)
 	ft_printf("%s:\n", filename);
 	ft_printf("total: %llu\n", max->blk);
 }
-
-/*
-**static void	get_dir_entries(DIR *dir, char *dir_name, t_mlist **entries)
-**{
-**	//int				status;
-**	struct dirent	*d_d;
-**	long			*acc;
-**	t_fileinfo		file;
-**	char			*dir_name_slash;
-**
-**	d_d = NULL;
-**	dir_name_slash = ft_strjoin(dir_name, "/");
-**	while ((d_d = readdir(dir)) && dir->__dd_loc < 4080)
-**	{
-**		get_fileinfo(&file, dir_name_slash, d_d->d_name);
-**		ft_mlstpush(entries, ft_mlstnew(&file, sizeof(t_fileinfo)));
-**	}
-**	if (dir->__dd_loc == 4080)
-**	{
-**		get_fileinfo(&file, dir_name_slash, d_d->d_name);
-**		ft_mlstpush(entries, ft_mlstnew(&file, sizeof(t_fileinfo)));
-**	}
-**	ft_strdel(&dir_name_slash);
-**}
-*/
 
 static void	traverce_subdir(size_t len, char *dir_name,
 							char *file_name)
